Rename to correct.cpp
